+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 [Unreleased]## [0.5.5.2b1]
+### Changed
+- Support enable_refine_conflict in Core02OptimizationEngine and ScenarioRunner
 
 ## [0.5.5.2b0]- 2023-05-02
 ### Added